+++ app.ts
@@ -5,6 +5,8 @@ import { errorHandler } from './src/middlewares/error-handler';
 import swaggerUI from 'swagger-ui-express';
 import { swDocument } from './swagger.def';
 import UserRouter from './src/routes/user.router';
+import MessagesRouter from './src/routes/message.router';
+import ConversationRouter from './src/routes/conversation.router';
 
 const app = Express();
 app.use(Express.urlencoded({ extended: true }));
@@ -13,6 +15,8 @@ app.use(corsHandler());
 app.use(typeCase('camel'));
 app.use('/api-docs', swaggerUI.serve, swaggerUI.setup(swDocument));
 app.use('/users', UserRouter);
+app.use('/messages', MessagesRouter);
+app.use('/conversations', ConversationRouter);
 app.use(errorHandler);
 
 module.exports = app;