+++ src/schemas/conversation.ts
@@ -1,0 +1,14 @@
+import { Joi } from '../services/validation';
+
+export const createConversationResquestSchema = Joi.object().keys({
+    userId: Joi.string(),
+});
+
+export const joinConversationResquestSchema = Joi.object().keys({
+  userId: Joi.string(),
+  conversationId: Joi.string(),
+});
+
+export const deleteConversationResquestSchema = Joi.object().keys({
+  id: Joi.string(),
+});